split misc into right test files
