+++ packages/ethereum-contracts/testsuites/all-contracts.js
@@ -1,16 +1,20 @@
-// TODO: deprecate this, and move them to the right categories
-require("../test/contracts/misc.test.js");
+// Upgradability
+require("../test/upgradability/UUPS.test.js");
 
+// Libs
+require("../test/libs/CallUtils.test.js");
+require("../test/libs/Int96SafeMath.test.js");
+require("../test/libs/UInt128SafeMath.test.js");
+
+// Superfluid
 require("./superfluid-core.js");
-
 require("./superfluid-governance.js");
-
 require("./custom-tokens.js");
 
-// apps helper contracts
+// App libraries
 require("./apps-contracts.js");
 
-// scenarios
+// Scenarios
 require("../test/scenarios/scenarios.js");
 
 // Utility contracts
@@ -18,4 +22,5 @@ require("../test/contracts/utils/BatchLiquidator.test.js");
 require("../test/contracts/utils/TOGA.test");
 
 // UX contracts
+require("../test/contracts/ux/Resolver.test.js");
 require("../test/contracts/ux/SuperUpgrader.test.js");
